+++ PABP_Projekat2/Controllers/ProductsController.cs
@@ -19,15 +19,26 @@ namespace PABP_Projekat2.Controllers
         }
 
         // GET: Products
-        public async Task<IActionResult> Index(int? categoryId, int? supplierId)
+        public async Task<IActionResult> Index(int categoryId = -1, int supplierId = -1)
         {
             //var northwindContext1 = _context.Products.Include(p => p.Category).Include(p => p.Supplier);
             var northwindContext = await _context.Products.ToListAsync();
-            if (categoryId != null && supplierId != null)
-                northwindContext = northwindContext.Where(p=> p.CategoryId == categoryId && p.SupplierId == supplierId).ToList();
 
-            ViewData["CategoryId"] = new SelectList(_context.Categories, "CategoryId", "CategoryName");
-            ViewData["SupplierId"] = new SelectList(_context.Suppliers, "SupplierId", "CompanyName");
+          
+            northwindContext = northwindContext.Where(p => p.CategoryId == (categoryId == -1 ? p.CategoryId : categoryId) && p.SupplierId == (supplierId == -1 ? p.SupplierId : supplierId)).ToList();
+          
+
+            var categoryIds = await _context.Categories.ToListAsync();
+            categoryIds.Add(new Category{ CategoryId= -1, CategoryName = "All" });
+
+            var supplerIds = await _context.Suppliers.ToListAsync();
+            supplerIds.Add(new Supplier { SupplierId = -1, CompanyName= "All" });
+
+            categoryIds.Reverse();
+            supplerIds.Reverse();
+
+            ViewData["CategoryId"] = new SelectList(categoryIds, "CategoryId", "CategoryName", categoryId);
+            ViewData["SupplierId"] = new SelectList(supplerIds, "SupplierId", "CompanyName", supplierId);
 
             return View(northwindContext);
         }
@@ -159,9 +170,15 @@ namespace PABP_Projekat2.Controllers
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var product = await _context.Products.FindAsync(id);
-            _context.Products.Remove(product);
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
+
+            if((await _context.OrderDetails.ToListAsync()).Where(x => x.ProductId == id).Count() == 0)
+            {
+                _context.Products.Remove(product);
+                await _context.SaveChangesAsync();
+                return RedirectToAction(nameof(Index));
+            }
+            ViewBag.checkError = "First delete all order details that contains this product!";
+            return View("Delete", product);
         }
 
         private bool ProductExists(int id)
